Add the first version of the first boot setup scripts
These scripts will auto-expand the partition and setup swap, and then
disable themselves.

Note however that running the script is disabled by default in the
rc.local file. This is because the way these are intended to be fetched
and run, there needs to be one reboot without them being active,
once they have already been copied into place.